20241125 - Phase 4 : Enhancement Contract

+++ src/main/java/com/fision/dto/ItemDetailsRequestDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ItemDetailsRequestDto {
     private String itemName;
-    private Integer paymentQuantity;
+    private Double paymentQuantity;
 }
